authentication made through cookies
