+++ index.js
@@ -1,76 +1,49 @@
 const express = require('express');
 const helmet = require('helmet');
 const cors = require('cors');
-const bcrypt = require('bcryptjs');
-
-const db = require('./data/dbConfig.js');
-const Users = require('./users/users-model.js');
-
+const userRouter = require('./routes/user-router')
+const session = require('express-session');
 const server = express();
+
+const sessionConfig = {
+  name: 'ladygaga',
+  secret: 'make it a little long and keep it safe!',
+  cookie: {
+    maxAge: 1000 * 60 * 60, // you need it if the cookie is to survive !!
+    secure: false, // with secure, the cookie only gets set when https !!
+    httpOnly: false,
+  },
+  resave: false,
+  saveUninitialized: false,
+}
 
 server.use(helmet());
 server.use(express.json());
 server.use(cors());
+server.use(session(sessionConfig));
 
 server.get('/', (req, res) => {
   res.send("It's alive!");
 });
 
-server.post('/api/register', (req, res) => {
-  const hash = bcrypt.hashSync(req.body.password, 11);
-  const newUser = {
-    username: req.body.username,
-    password: hash
-  }
+server.use('/api/', userRouter);
 
-  Users.add(newUser)
-    .then(saved => {
-      res.status(201).json(saved);
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
 
-server.post('/api/login', (req, res) => {
-  let { username, password } = req.body;
 
-  Users.findBy({ username })
-    .first()
-    .then(user => {
-      if (user &&  bcrypt.compareSync(password, user.password)) {
-        res.status(200).json({ message: `Welcome ${user.username}!` });
-      } else {
-        res.status(401).json({ message: 'Invalid Credentials' });
-      }
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
+// function restricted(req, res, next) {
+//   const { username, password } = req.headers;
 
-server.get('/api/users', restricted, (req, res) => {
-  Users.find()
-    .then(users => {
-      res.json(users);
-    })
-    .catch(err => res.send(err));
-});
-
-function restricted(req, res, next) {
-  const { username, password } = req.headers;
-
-  Users.findBy({ username: username }).first()
-    .then(user => {
-      if(user && bcrypt.compareSync(password, user.password)){
-        next();
-      } else {
-        res.status(401).json({ message: 'invalid credentials ' })
-      }
-    }).catch(err => {
-      res.status(401).json({ message: err.message })
-    })
-}
+//   Users.findBy({ username: username }).first()
+//     .then(user => {
+//       if(user && bcrypt.compareSync(password, user.password)){
+//         next();
+//       } else {
+//         res.status(401).json({ message: 'invalid credentials ' })
+//       }
+//     }).catch(err => {
+//       res.status(401).json({ message: err.message })
+//     })
+// }
 
 
 const port = process.env.PORT || 5000;
