+++ middlewares/session-checker.js
@@ -1,0 +1,7 @@
+module.exports = (req, res, next) => {
+  if (req.session && req.session.user) {
+    next();
+  } else {
+    res.status(400).json({ message: 'No credentials provided' });
+  }
+};
